chore(common): split level store import path
