+++ .changeset/split-level-store.md
@@ -1,0 +1,6 @@
+---
+"@enbox/common": patch
+"@enbox/agent": patch
+---
+
+chore: move the Level-backed common store behind a dedicated optional subpath.
